+++ lib/preface_core/void.ml
@@ -1,0 +1,7 @@
+type t = |
+
+let absurd : t -> 'a = function _ -> .
+
+let left x = Either.fold ~left:Fun.id ~right:absurd x
+
+let right x = Either.fold ~left:absurd ~right:Fun.id x