+++ lib/preface_specs/divisible.mli
@@ -1,6 +1,24 @@
 (** [Divisible] is a "Contravariant Applicative Functor", in other word,
     [Divisible] is the dual of an {!module:Applicative} *)
 
+(** {2 Laws}
+
+    To have a predictable behaviour, the instance of [Divisible] must obey some
+    laws.
+
+    + All {!module:Contravariant} laws
+    + [divide f x conquer = contramap (Stdlib.fst % f) x]
+    + [divide f conquer x = contramap (Stdlib.snd % f) x]
+    + {[
+        divide f (divide g m n) o
+        = divide
+            (fun x ->
+              let (bc, _) = f x in
+              let (b, c) = g bc in
+              (a, (b, c)))
+            m (divide id n o)
+      ]} *)
+
 (** {1 Minimal definition} *)
 
 (** Exposes the [divide] and [conquer] functions, mandatory for each
@@ -43,6 +61,15 @@ end
 module type INFIX = sig
   type 'a t
   (** The type held by the [Divisible]. *)
+
+  val ( >*< ) : 'a t -> 'b t -> ('a * 'b) t
+  (** Infix version of [divided]. *)
+
+  val ( >* ) : 'a t -> unit t -> 'a t
+  (** Discard the value of the second argument. *)
+
+  val ( *< ) : unit t -> 'a t -> 'a t
+  (** Discard the value of the first argument. *)
 
   include Contravariant.INFIX with type 'a t := 'a t
 end