+++ lib/preface_stdlib/void.mli
@@ -1,0 +1,2 @@
+include module type of Preface_core.Void
+(** @inline *)